% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eye_Contact.R
\docType{data}
\name{Eye_Contact}
\alias{Eye_Contact}
\title{Eye_Contact}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Eye_Contact
}
\description{
A data set containing recorded data of eye contact the "subject" made.
}
\keyword{datasets}
