% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gesture.R
\docType{data}
\name{Gesture}
\alias{Gesture}
\title{Gesture}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Gesture
}
\description{
A data set containing a reaction of subjects with relations to food and human's gestures.
}
\keyword{datasets}
