% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inhibitory_Control.R
\docType{data}
\name{Inhibitory_Control}
\alias{Inhibitory_Control}
\title{Inhibitory_Control}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Inhibitory_Control
}
\description{
A data set containing recorded data of food route trials.
}
\keyword{datasets}
