% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\docType{data}
\name{Summary}
\alias{Summary}
\title{Summary}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Summary
}
\description{
The data set contains an overall summary data of the other sets of data.
}
\keyword{datasets}
