% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Temperament.R
\docType{data}
\name{Temperament}
\alias{Temperament}
\title{Temperament}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Temperament
}
\description{
A data set containing data on subjects interactions among objects
and individuals.
}
\keyword{datasets}
