% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Working_memory.R
\docType{data}
\name{Working_memory}
\alias{Working_memory}
\title{Working_memory}
\format{
A data frame with X number of columns and Y number of rows
}
\source{

}
\usage{
Working_memory
}
\description{
A data set containing subjects memory test performance.
}
\keyword{datasets}
